Refactor components (#563)
* finished reworking initial components

* reworked components a bit

* fixed cached data setting

* ran format

* reformatted components section

* reverted attention changes

* reverted layer norm changes

* reverted mlp changes

* Revert "reverted attention changes"

This reverts commit 91a5712194563f35766d8946facd4f53ba2a2f84.

* Revert "reverted layer norm changes"

This reverts commit cecf93ed4204fc316875584a7024a4302035c4c2.

* Revert "reverted mlp changes"

This reverts commit 9d8e91a244247be411050c3dc8c1c0c02f092abd.

* removed some model tests

* Revert "removed some model tests"

This reverts commit ef4518b3e848d799409c8b7045ec7206dc68bf23.

* removed model tests

* added model back

* lowered accuracy

* Revert "lowered accuracy"

This reverts commit beb014e7051ce08f545b427aa58c484c581525e4.

* added model back to test loop

* reverted accuracy change

* added proper headers

* added import again

* updated attention component

* added new line

* removed import

* cleaned imports

* ran black

* updated doc string

* finished fixing docstring

* fixed mypi error
