+++ transformer_lens/components/rms_norm_pre.py
@@ -1,0 +1,37 @@
+"""Hooked Transformer RMS Norm Pre Component.
+
+This module contains all the component :class:`RMSNormPre`.
+"""
+from typing import Dict, Union
+
+import torch
+import torch.nn as nn
+from jaxtyping import Float
+
+from transformer_lens.hook_points import HookPoint
+from transformer_lens.HookedTransformerConfig import HookedTransformerConfig
+
+
+class RMSNormPre(nn.Module):
+    def __init__(self, cfg: Union[Dict, HookedTransformerConfig]):
+        """RMSNormPre - LayerNormPre without the centering and bias (RMS = Root Mean Square)"""
+        super().__init__()
+        if isinstance(cfg, Dict):
+            cfg = HookedTransformerConfig.from_dict(cfg)
+        self.cfg = cfg
+        self.eps = self.cfg.eps
+
+        # Adds a hook point for the normalisation scale factor
+        self.hook_scale = HookPoint()  # [batch, pos]
+        self.hook_normalized = HookPoint()  # [batch, pos, length]
+
+    def forward(
+        self, x: Float[torch.Tensor, "batch pos length"]
+    ) -> Float[torch.Tensor, "batch pos length"]:
+        if self.cfg.dtype not in [torch.float32, torch.float64]:
+            x = x.to(torch.float32)
+
+        scale: Float[torch.Tensor, "batch pos 1"] = self.hook_scale(
+            (x.pow(2).mean(-1, keepdim=True) + self.eps).sqrt()
+        )
+        return self.hook_normalized(x / scale).to(self.cfg.dtype)  # [batch, pos, length]
